fix(app): Detect main window changes on wm-triggered activations
Previously we did not check for main window changed when handling
activation events that we knew were initiated by us. Maybe this worked
in a previous version of macOS but it doesn't anymore; make_key_window
causes the activation event to be sent ahead of the main window changed
event.