fix(#10): Use skylight APIs for backup window destroy events
This API is apparently more reliable. The  problem is that events still
seem to get delayed until something else happens; sometimes a different
event or even user input seems to trigger it. I don't know why yet. In
practice it should be a pretty small issue, and we still use the AX
events if those are available, so the impact should be small.

Fixes #10.